Style and refactor scroll to bottom button into separate view
