+++ OLMoE.swift/Constants/Localizable.xcstrings
@@ -20,6 +20,9 @@
     "%d total" : {
 
     },
+    "%lld" : {
+
+    },
     "%lld%%" : {
 
     },
